+++ tests/server/app/test_processing.py
@@ -19,30 +19,21 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 import json
-import os
 import pkgutil
-import time
 from typing import Dict
 
 import yaml
 from xcube.constants import EXTENSION_POINT_SERVER_APIS
-from xcube.server.testing import ServerTest
 from xcube.util import extension
 from xcube.util.extension import ExtensionRegistry
 
-from . import test_utils
 from xcube_geodb_openeo.backend import processes
 from xcube_geodb_openeo.backend.processes import LoadCollection
+from . import test_utils
+from .base_test import BaseTest
 
 
-class ProcessingTest(ServerTest):
-
-    def setUp(self) -> None:
-        super().setUp()
-        wait_for_server_startup = os.environ.get('WAIT_FOR_STARTUP',
-                                                 '0') == '1'
-        if wait_for_server_startup:
-            time.sleep(10)
+class ProcessingTest(BaseTest):
 
     def add_extension(self, er: ExtensionRegistry) -> None:
         er.add_extension(
